requirements.txt
aiomqtt==2.3.0
asyncio==3.4.3
flatdict==4.0.1
mypy==1.15.0
mypy-extensions==1.0.0
orjson==3.10.15
paho-mqtt==2.1.0
python-dotenv==1.0.1
PyYAML==6.0.2
setuptools==75.8.0
typing_extensions==4.12.2
wheel==0.45.1
